requirements.txt
Flask==2.3.3
flask-restx==1.3.0
google-cloud-storage==2.10.0
requests==2.31.0
python-dateutil==2.8.2
gunicorn==21.2.0
pytest==7.4.2
pytest-cov==4.1.0
